+++ CHANGELOG.md
@@ -1,3 +1,18 @@
+## develop
+
+- Add `LoggingNotification` class.
+- Add `LoggingLevel` enum.
+- Add `ProtocolHandler.sendNotification` method.
+- Add `ProtocolHandler.sendLoggingNotification` method.
+- Add `withArray` function.
+- Add `withObject` function.
+- Add `name` property to `CallToolRequest`.
+- Remove `NotificationParams` class.
+- Restruct `JsonRpcNotification` class.
+- `Tool` constructor validates `name`.
+- Rename `Tool.parameters` to `Tool.inputSchema`.
+- Remove `ProtocolHandler.sendNotificationToClient` method.
+
 ## 0.5.0
 
 - Initial version.